nccl_ep: fix int32 overflow in HT slot offsets
Problem: The HT dispatch/combine JIT kernels compute buffer offsets as
slot_index * stride with both operands int32, then add the product to a
typed pointer. The receiver slot space is partitioned by source rank at
max_tokens_per_rank granularity, so with 32 LSA ranks x
max_tokens_per_rank=32768 x hidden_dim=4096 the dispatch destination
offset crosses 2^31 for sender ranks >= 16, producing a negative offset
and a deterministic CUDA_ERROR_ILLEGAL_ADDRESS in
dispatch_nodes1_lsa32_* on the upper half of the ranks (observed on a
397B-parameter MoE at TP4/EP32: ranks 20-31 fault, the fault follows
the rank index across hosts, and is independent of the capacity
factor; max_tokens_per_rank=16384 runs clean with zero margin --
524287*4096 is 4 KiB below 2^31). The combine-side read at
slot * HIDDEN_DIM * size_u16 is in bytes and overflows at half the
token count; the local-dup zone write scales with the full receive
capacity.

Solution: Cast the slot index to size_t at every slot-scale multiply in
dispatch_s2g_issue_token (token/prob/scaling-factor destinations), the
combine g2s source reads, and the local-dup pad writes. Header-only
change to the runtime-JIT'd device source; no library binary is
affected and the JIT cache re-keys on the new source automatically.

Limitations: Only the slot-scale multiplies that can exceed 2^31 under
currently supported configurations are widened; other index arithmetic
bounded by per-chunk or per-warp extents remains int32.

+++ contrib/nccl_ep/device/hybrid_ep.cuh
@@ -1397,15 +1397,15 @@ __forceinline__ __device__ void dispatch_s2g_issue_token(
     const uint32_t sf_bytes = (uint32_t)sf_bytes_per_token;
 
     // Remote fan-out: each field has its own output array, indexed [rank]+slot*stride.
-    const void* token_dst = remote_expert_output_token[dst.remote_rank_id] + (dst.output_buffer_index * HIDDEN_DIM);
+    const void* token_dst = remote_expert_output_token[dst.remote_rank_id] + ((size_t)dst.output_buffer_index * HIDDEN_DIM);
     const void* prob_dst = nullptr;
     const void* sf_dst = nullptr;
     if constexpr (FORWARD_DISPATCH) {
         prob_dst = remote_expert_output_prob[dst.remote_rank_id] +
-                   (dst.output_buffer_index * (experts_per_rank * LSA_TEAM_SIZE));
+                   ((size_t)dst.output_buffer_index * (experts_per_rank * LSA_TEAM_SIZE));
     }
     if constexpr (HAS_SF) {
-        sf_dst = remote_expert_output_scaling_factor[dst.remote_rank_id] + dst.output_buffer_index * sf_bytes_per_token;
+        sf_dst = remote_expert_output_scaling_factor[dst.remote_rank_id] + (size_t)dst.output_buffer_index * sf_bytes_per_token;
     }
     copy_token_bundle<TOKEN_DATA_TYPE, SMEM_TYPE, FORWARD_DISPATCH, HAS_SF, copy_dir::to_gmem>(
         smem_buffer_ptr,
@@ -2123,10 +2123,10 @@ __forceinline__ __device__ void issue_local_g2s_row(
             slot = pending_s2d;
         }
         const uint16_t* token_src =
-            remote_expert_input_token[rank_id] + (slot * HIDDEN_DIM * nccl_ep::size_u16<kTokenDtype>());
+            remote_expert_input_token[rank_id] + ((size_t)slot * HIDDEN_DIM * nccl_ep::size_u16<kTokenDtype>());
         const float* prob_src = nullptr;
         if constexpr (BACKWARD_COMBINE) {
-            prob_src = remote_expert_input_prob[rank_id] + (slot * (experts_per_rank * num_of_ranks_per_node));
+            prob_src = remote_expert_input_prob[rank_id] + ((size_t)slot * (experts_per_rank * num_of_ranks_per_node));
         }
         issue_g2s_entry<INTER_NODE, BACKWARD_COMBINE, /*WRITE_LAST_FLAG=*/true>(
             smem_buffer_ptr,
@@ -3782,7 +3782,7 @@ __forceinline__ __device__ void PAD_warp_group_device_function(
         int32_t pad = rem ? (alignment - rem) : 0;
         for (int p = 0; p < pad; p++, row_idx++) {
             if ((row_idx % global_stride) == global_id) {
-                void* dst = reinterpret_cast<void*>(local_buf + (zone_offsets[e] + count + p) * hidden_dim);
+                void* dst = reinterpret_cast<void*>(local_buf + (size_t)(zone_offsets[e] + count + p) * hidden_dim);
                 cuda::ptx::cp_async_bulk(
                     cuda::ptx::space_global,
                     cuda::ptx::space_shared,
